+++ src/modelo/Pasaje.java
@@ -4,6 +4,8 @@
  */
 package modelo;
 
+import java.math.BigDecimal;
+
 /**
  *
  * @author aguse
@@ -11,12 +13,12 @@ package modelo;
 public class Pasaje {
     private int id_pasaje;
     private String tipoTransporte;
-    private double importe;
+    private BigDecimal importe;
     private Ciudad ciudadOrigen;
     private Ciudad ciudadDestino;
     private boolean estado;
 
-    public Pasaje(int id_pasaje, String tipoTransporte, double importe, Ciudad ciudadOrigen, Ciudad ciudadDestino, boolean estado) {
+    public Pasaje(int id_pasaje, String tipoTransporte, BigDecimal importe, Ciudad ciudadOrigen, Ciudad ciudadDestino, boolean estado) {
         this.id_pasaje = id_pasaje;
         this.tipoTransporte = tipoTransporte;
         this.importe = importe;
@@ -25,7 +27,7 @@ public class Pasaje {
         this.estado = estado;
     }
 
-    public Pasaje(String tipoTransporte, double importe, Ciudad ciudadOrigen, Ciudad ciudadDestino, boolean estado) {
+    public Pasaje(String tipoTransporte, BigDecimal importe, Ciudad ciudadOrigen, Ciudad ciudadDestino, boolean estado) {
         this.tipoTransporte = tipoTransporte;
         this.importe = importe;
         this.ciudadOrigen = ciudadOrigen;
@@ -52,11 +54,11 @@ public class Pasaje {
         this.tipoTransporte = tipoTransporte;
     }
 
-    public double getImporte() {
+    public BigDecimal getImporte() {
         return importe;
     }
 
-    public void setImporte(double importe) {
+    public void setImporte(BigDecimal importe) {
         this.importe = importe;
     }
 
@@ -89,5 +91,5 @@ public class Pasaje {
         return "Pasaje{" + "id_pasaje=" + id_pasaje + ", tipoTransporte=" + tipoTransporte + ", importe=" + importe + ", ciudadOrigen=" + ciudadOrigen + ", ciudadDestino=" + ciudadDestino + ", estado=" + estado + '}';
     }
 
-   
+    
 }
